ajustes na aplicação para tratamento de !!

+++ src/main/kotlin/com/samuca/todolist/controller/PreferenciasRestController.kt
@@ -18,25 +18,32 @@ class PreferenciasRestController(private val preferenciasDao: PreferenciasDao) {
         val authentication: Authentication = SecurityContextHolder.getContext().authentication
         val (id) = authentication.getPrincipal() as Usuario
 
-        val preferencias: Preferencias? = preferenciasDao.findByUsuarioId(id!!)
+        if (id != null) {
+            val preferencias: Preferencias? = preferenciasDao.findByUsuarioId(id)
+            return preferencias
+        }
 
-        return preferencias
+        return null
     }
 
     @PostMapping("/preferencias")
-    fun save(@RequestBody @Valid preferencias: Preferencias): Preferencias {
+    fun save(@RequestBody @Valid preferencias: Preferencias): Preferencias? {
         val authentication: Authentication = SecurityContextHolder.getContext().authentication
         val (id) = authentication.getPrincipal() as Usuario
 
-        val preferenciasExistente: Preferencias? = preferenciasDao.findByUsuarioId(id!!)
-        if (preferenciasExistente != null) {
-            preferencias.id = preferenciasExistente.id
+        if (id != null) {
+            val preferenciasExistente: Preferencias? = preferenciasDao.findByUsuarioId(id)
+            if (preferenciasExistente != null) {
+                preferencias.id = preferenciasExistente.id
+            }
+
+            preferencias.usuarioId = id
+
+            val retorno: Preferencias = preferenciasDao.save(preferencias)
+            return retorno
         }
 
-        preferencias.usuarioId = id!!
-
-        val retorno: Preferencias = preferenciasDao.save(preferencias)
-        return retorno
+        return null
     }
 
     @DeleteMapping("/preferencia")
@@ -44,14 +51,18 @@ class PreferenciasRestController(private val preferenciasDao: PreferenciasDao) {
         val authentication: Authentication = SecurityContextHolder.getContext().authentication
         val (id) = authentication.getPrincipal() as Usuario
 
-        val preferencias: Preferencias? = preferenciasDao.findByUsuarioId(id!!)
-        if (preferencias == null) {
-            return ResponseEntity.badRequest().body("Nao existe preferencias para o usuario")
+        if (id != null) {
+            val preferencias: Preferencias? = preferenciasDao.findByUsuarioId(id)
+            if (preferencias == null) {
+                return ResponseEntity.badRequest().body("Nao existe preferencias para o usuario")
+            }
+
+            preferenciasDao.delete(preferencias)
+
+            return ResponseEntity.ok("OK")
         }
 
-        preferenciasDao.delete(preferencias)
-
-        return ResponseEntity.ok("OK")
+        return ResponseEntity.badRequest().body("Nao existe tarefa para o usuario logado ")
     }
 
 }
